+++ SharpCoreSource/SharpCore/src/Ticking/Data/TickSimulationConfigData.cs
@@ -2,7 +2,7 @@ namespace SharpCore.Ticking
 {
     public class TickSimulationConfigData : TickBaseConfigData
     {
-        public double tickrate;
-        public double maxDelta;
+        public float tickrate;
+        public float maxDelta;
     }
 }
