+++ SharpCoreSource/SharpCoreTests/Ticking/Demo/DemoRenderTickClientInstanceRegistrationTest.cs
@@ -28,7 +28,7 @@ namespace SharpCoreTests.Ticking.Demo
         
         #region Tick
 
-        void ITickRenderClient.Tick(double delta)
+        void ITickRenderClient.Tick(float delta)
         {
             
         }
